Fix seed file

+++ db/seeds.rb
@@ -1,28 +1,33 @@
-# This file should contain all the record creation needed to seed the database with its default values.
-# The data can then be loaded with the rake db:seed (or created alongside the db with db:setup).
-#
-# Examples:
-#
-#   cities = City.create([{ name: 'Chicago' }, { name: 'Copenhagen' }])
-#   Mayor.create(name: 'Emanuel', city: cities.first)
+# frozen_string_literal: true
 
-puts "This will not touch any users or projects..."
-puts "This should be run after making the Tetris project and any other games that have scores..."
-puts "Destroying all scores, levels and lines..."
+return unless Rails.env.development?
+
+puts 'Creating project Tetris'
+Project.find_or_create_by(
+  title: 'tetris',
+  private: false
+)
+
+puts 'Destroying all scores, levels and lines...'
 Score.destroy_all
 
-puts "Creating 5 scores, names, levels and lines for Tetris"
-5.times do |number|
-  score = Project.where(:title => "Tetris").first.scores.new
-  score.score = number
-  score.name = "Chubb"
-  score.save
-
-  level = Level.new
-  level.level = number
-  score.level = level
-
-  line = Line.new
-  line.lines = number
-  score.line = line
+puts 'Creating scores for Tetris'
+NUMBER_OF_SCORES = 5
+NUMBER_OF_SCORES.times do |n|
+  score = Project.find_by(title: 'tetris').scores.create(
+    score: n,
+    name: 'Chubb'
+  )
+  level = Level.create(
+    level: n,
+    score: score
+  )
+  line = Line.create(
+    lines: n,
+    score: score
+  )
+  score.update!(
+    level: level,
+    line: line
+  )
 end
